Added comments to actions page and moved it to its own folder.
